v1.0.4: - Ditched XInput in favor of the game's internal controller logic (DualSense is supported now) - Autoplay is working again. The music player now deactivates in MULE and BT zones - No longer using DirectXHook. This potentially fixed a rare issue where DXH would crash on startup - No longer using imgui. This was only used to show a progress bar when the mod loads - Fixed a bug where the music player UI would show when using watchtowers

+++ MusicMod/include/InputCode.h
@@ -8,7 +8,7 @@ enum class InputSource
 
 struct InputCode
 {
-	int code;
+	uint64_t code;
 	InputSource source;
 
 	bool operator==(const InputCode& other) const {
